0.5.32

+++ src/Notio.Common/Constants/Timeouts.cs
@@ -8,30 +8,30 @@ public static class Timeouts
     /// <summary>
     /// For ultra-fast control packets (e.g., ping, small config).
     /// </summary>
-    public const int Instant = 500;
+    public const ushort Instant = 500;
 
     /// <summary>
     /// For lightweight operations such as setting connection options.
     /// </summary>
-    public const int Short = 1000;
+    public const ushort Short = 1000;
 
     /// <summary>
     /// Default timeout for general-purpose packets.
     /// </summary>
-    public const int Default = 5000;
+    public const ushort Default = 5000;
 
     /// <summary>
     /// For operations that may take moderate time (e.g., compression, key exchange).
     /// </summary>
-    public const int Moderate = 8000;
+    public const ushort Moderate = 8000;
 
     /// <summary>
     /// For long-running tasks (e.g., database query, file processing).
     /// </summary>
-    public const int Long = 15000;
+    public const ushort Long = 15000;
 
     /// <summary>
     /// For very long-running operations (e.g., file upload, sync).
     /// </summary>
-    public const int Extended = 30000;
+    public const ushort Extended = 30000;
 }
